Upgrade to semantic convention 1.25.0
