+++ api/include/opentelemetry/trace/semantic_conventions.h
@@ -22,76 +22,14 @@ namespace SemanticConventions
 /**
  * The URL of the OpenTelemetry schema for these keys and values.
  */
-static constexpr const char *kSchemaUrl = "https://opentelemetry.io/schemas/1.24.0";
-
-/**
- * The name of the invoked function.
- *
- * <p>Notes:
-  <ul> <li>SHOULD be equal to the {@code faas.name} resource attribute of the invoked function.</li>
- </ul>
- */
-static constexpr const char *kFaasInvokedName = "faas.invoked_name";
-
-/**
- * The cloud provider of the invoked function.
- *
- * <p>Notes:
-  <ul> <li>SHOULD be equal to the {@code cloud.provider} resource attribute of the invoked
- function.</li> </ul>
- */
-static constexpr const char *kFaasInvokedProvider = "faas.invoked_provider";
-
-/**
- * The cloud region of the invoked function.
- *
- * <p>Notes:
-  <ul> <li>SHOULD be equal to the {@code cloud.region} resource attribute of the invoked
- function.</li> </ul>
- */
-static constexpr const char *kFaasInvokedRegion = "faas.invoked_region";
-
-/**
- * Type of the trigger which caused this function invocation.
- */
-static constexpr const char *kFaasTrigger = "faas.trigger";
-
-/**
- * The <a href="/docs/resource/README.md#service">{@code service.name}</a> of the remote service.
- * SHOULD be equal to the actual {@code service.name} resource attribute of the remote service if
- * any.
- */
-static constexpr const char *kPeerService = "peer.service";
-
-/**
- * Username or client_id extracted from the access token or <a
- * href="https://tools.ietf.org/html/rfc7235#section-4.2">Authorization</a> header in the inbound
- * request from outside the system.
- */
-static constexpr const char *kEnduserId = "enduser.id";
-
-/**
- * Actual/assumed role the client is making the request under extracted from token or application
- * security context.
- */
-static constexpr const char *kEnduserRole = "enduser.role";
-
-/**
- * Scopes or granted authorities the client currently possesses extracted from token or application
- * security context. The value would come from the scope associated with an <a
- * href="https://tools.ietf.org/html/rfc6749#section-3.3">OAuth 2.0 Access Token</a> or an attribute
- * value in a <a
- * href="http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-tech-overview-2.0.html">SAML 2.0
- * Assertion</a>.
- */
-static constexpr const char *kEnduserScope = "enduser.scope";
+static constexpr const char *kSchemaUrl = "https://opentelemetry.io/schemas/1.25.0";
 
 /**
  * Identifies the class / type of event.
  *
  * <p>Notes:
   <ul> <li>Event names are subject to the same rules as <a
- href="https://github.com/open-telemetry/opentelemetry-specification/tree/v1.26.0/specification/common/attribute-naming.md">attribute
+ href="https://github.com/open-telemetry/opentelemetry-specification/tree/v1.31.0/specification/common/attribute-naming.md">attribute
  names</a>. Notably, event names are namespaced to avoid collisions and provide a clean separation
  of semantics for events in separate domains like browser, mobile, and kubernetes.</li> </ul>
  */
@@ -157,9 +95,9 @@ static constexpr const char *kAndroidState = "android.state";
 
 /**
  * The name of the connection pool; unique within the instrumented application. In case the
- * connection pool implementation doesn't provide a name, then the <a
- * href="/docs/database/database-spans.md#connection-level-attributes">db.connection_string</a>
- * should be used
+ * connection pool implementation doesn't provide a name, instrumentation should use a combination
+ * of {@code server.address} and {@code server.port} attributes formatted as {@code
+ * server.address:server.port}.
  */
 static constexpr const char *kPoolName = "pool.name";
 
@@ -167,6 +105,11 @@ static constexpr const char *kPoolName = "pool.name";
  * The state of a connection in the pool
  */
 static constexpr const char *kState = "state";
+
+/**
+ * Rate-limiting result, shows whether the lease was acquired or contains a rejection reason
+ */
+static constexpr const char *kAspnetcoreRateLimitingResult = "aspnetcore.rate_limiting.result";
 
 /**
  * Full type name of the <a
@@ -180,11 +123,6 @@ static constexpr const char *kAspnetcoreDiagnosticsHandlerType =
  * Rate limiting policy name.
  */
 static constexpr const char *kAspnetcoreRateLimitingPolicy = "aspnetcore.rate_limiting.policy";
-
-/**
- * Rate-limiting result, shows whether the lease was acquired or contains a rejection reason
- */
-static constexpr const char *kAspnetcoreRateLimitingResult = "aspnetcore.rate_limiting.result";
 
 /**
  * Flag indicating if request was handled by the application pipeline.
@@ -234,6 +172,12 @@ static constexpr const char *kJvmMemoryPoolName = "jvm.memory.pool.name";
 static constexpr const char *kJvmMemoryType = "jvm.memory.type";
 
 /**
+ * The CPU state for this data point. A process SHOULD be characterized <em>either</em> by data
+ * points with no {@code state} labels, <em>or only</em> data points with {@code state} labels.
+ */
+static constexpr const char *kProcessCpuState = "process.cpu.state";
+
+/**
  * The device identifier
  */
 static constexpr const char *kSystemDevice = "system.device";
@@ -244,7 +188,8 @@ static constexpr const char *kSystemDevice = "system.device";
 static constexpr const char *kSystemCpuLogicalNumber = "system.cpu.logical_number";
 
 /**
- * The state of the CPU
+ * The CPU state for this data point. A system's CPU SHOULD be characterized <em>either</em> by data
+ * points with no {@code state} labels, <em>or only</em> data points with {@code state} labels.
  */
 static constexpr const char *kSystemCpuState = "system.cpu.state";
 
@@ -298,7 +243,177 @@ static constexpr const char *kSystemNetworkState = "system.network.state";
  * href="https://man7.org/linux/man-pages/man1/ps.1.html#PROCESS_STATE_CODES">Linux Process State
  * Codes</a>
  */
-static constexpr const char *kSystemProcessesStatus = "system.processes.status";
+static constexpr const char *kSystemProcessStatus = "system.process.status";
+
+/**
+ * Uniquely identifies the framework API revision offered by a version ({@code os.version}) of the
+ * android operating system. More information can be found <a
+ * href="https://developer.android.com/guide/topics/manifest/uses-sdk-element#ApiLevels">here</a>.
+ */
+static constexpr const char *kAndroidOsApiLevel = "android.os.api_level";
+
+/**
+ * The JSON-serialized value of each item in the {@code AttributeDefinitions} request field.
+ */
+static constexpr const char *kAwsDynamodbAttributeDefinitions =
+    "aws.dynamodb.attribute_definitions";
+
+/**
+ * The value of the {@code AttributesToGet} request parameter.
+ */
+static constexpr const char *kAwsDynamodbAttributesToGet = "aws.dynamodb.attributes_to_get";
+
+/**
+ * The value of the {@code ConsistentRead} request parameter.
+ */
+static constexpr const char *kAwsDynamodbConsistentRead = "aws.dynamodb.consistent_read";
+
+/**
+ * The JSON-serialized value of each item in the {@code ConsumedCapacity} response field.
+ */
+static constexpr const char *kAwsDynamodbConsumedCapacity = "aws.dynamodb.consumed_capacity";
+
+/**
+ * The value of the {@code Count} response parameter.
+ */
+static constexpr const char *kAwsDynamodbCount = "aws.dynamodb.count";
+
+/**
+ * The value of the {@code ExclusiveStartTableName} request parameter.
+ */
+static constexpr const char *kAwsDynamodbExclusiveStartTable = "aws.dynamodb.exclusive_start_table";
+
+/**
+ * The JSON-serialized value of each item in the {@code GlobalSecondaryIndexUpdates} request field.
+ */
+static constexpr const char *kAwsDynamodbGlobalSecondaryIndexUpdates =
+    "aws.dynamodb.global_secondary_index_updates";
+
+/**
+ * The JSON-serialized value of each item of the {@code GlobalSecondaryIndexes} request field
+ */
+static constexpr const char *kAwsDynamodbGlobalSecondaryIndexes =
+    "aws.dynamodb.global_secondary_indexes";
+
+/**
+ * The value of the {@code IndexName} request parameter.
+ */
+static constexpr const char *kAwsDynamodbIndexName = "aws.dynamodb.index_name";
+
+/**
+ * The JSON-serialized value of the {@code ItemCollectionMetrics} response field.
+ */
+static constexpr const char *kAwsDynamodbItemCollectionMetrics =
+    "aws.dynamodb.item_collection_metrics";
+
+/**
+ * The value of the {@code Limit} request parameter.
+ */
+static constexpr const char *kAwsDynamodbLimit = "aws.dynamodb.limit";
+
+/**
+ * The JSON-serialized value of each item of the {@code LocalSecondaryIndexes} request field.
+ */
+static constexpr const char *kAwsDynamodbLocalSecondaryIndexes =
+    "aws.dynamodb.local_secondary_indexes";
+
+/**
+ * The value of the {@code ProjectionExpression} request parameter.
+ */
+static constexpr const char *kAwsDynamodbProjection = "aws.dynamodb.projection";
+
+/**
+ * The value of the {@code ProvisionedThroughput.ReadCapacityUnits} request parameter.
+ */
+static constexpr const char *kAwsDynamodbProvisionedReadCapacity =
+    "aws.dynamodb.provisioned_read_capacity";
+
+/**
+ * The value of the {@code ProvisionedThroughput.WriteCapacityUnits} request parameter.
+ */
+static constexpr const char *kAwsDynamodbProvisionedWriteCapacity =
+    "aws.dynamodb.provisioned_write_capacity";
+
+/**
+ * The value of the {@code ScanIndexForward} request parameter.
+ */
+static constexpr const char *kAwsDynamodbScanForward = "aws.dynamodb.scan_forward";
+
+/**
+ * The value of the {@code ScannedCount} response parameter.
+ */
+static constexpr const char *kAwsDynamodbScannedCount = "aws.dynamodb.scanned_count";
+
+/**
+ * The value of the {@code Segment} request parameter.
+ */
+static constexpr const char *kAwsDynamodbSegment = "aws.dynamodb.segment";
+
+/**
+ * The value of the {@code Select} request parameter.
+ */
+static constexpr const char *kAwsDynamodbSelect = "aws.dynamodb.select";
+
+/**
+ * The number of items in the {@code TableNames} response parameter.
+ */
+static constexpr const char *kAwsDynamodbTableCount = "aws.dynamodb.table_count";
+
+/**
+ * The keys in the {@code RequestItems} object field.
+ */
+static constexpr const char *kAwsDynamodbTableNames = "aws.dynamodb.table_names";
+
+/**
+ * The value of the {@code TotalSegments} request parameter.
+ */
+static constexpr const char *kAwsDynamodbTotalSegments = "aws.dynamodb.total_segments";
+
+/**
+ * Array of brand name and version separated by a space
+ *
+ * <p>Notes:
+  <ul> <li>This value is intended to be taken from the <a
+ href="https://wicg.github.io/ua-client-hints/#interface">UA client hints API</a> ({@code
+ navigator.userAgentData.brands}).</li> </ul>
+ */
+static constexpr const char *kBrowserBrands = "browser.brands";
+
+/**
+ * Preferred language of the user using the browser
+ *
+ * <p>Notes:
+  <ul> <li>This value is intended to be taken from the Navigator API {@code
+ navigator.language}.</li> </ul>
+ */
+static constexpr const char *kBrowserLanguage = "browser.language";
+
+/**
+ * A boolean that is true if the browser is running on a mobile device
+ *
+ * <p>Notes:
+  <ul> <li>This value is intended to be taken from the <a
+ href="https://wicg.github.io/ua-client-hints/#interface">UA client hints API</a> ({@code
+ navigator.userAgentData.mobile}). If unavailable, this attribute SHOULD be left unset.</li> </ul>
+ */
+static constexpr const char *kBrowserMobile = "browser.mobile";
+
+/**
+ * The platform on which the browser is running
+ *
+ * <p>Notes:
+  <ul> <li>This value is intended to be taken from the <a
+href="https://wicg.github.io/ua-client-hints/#interface">UA client hints API</a> ({@code
+navigator.userAgentData.platform}). If unavailable, the legacy {@code navigator.platform} API SHOULD
+NOT be used instead and this attribute SHOULD be left unset in order for the values to be
+consistent. The list of possible values is defined in the <a
+href="https://wicg.github.io/ua-client-hints/#sec-ch-ua-platform">W3C User-Agent Client Hints
+specification</a>. Note that some (but not all) of these values can overlap with values in the <a
+href="./os.md">{@code os.type} and {@code os.name} attributes</a>. However, for consistency, the
+values in the {@code browser.platform} attribute should capture the exact value that the user agent
+provides.</li> </ul>
+ */
+static constexpr const char *kBrowserPlatform = "browser.platform";
 
 /**
  * Client address - domain name if available without reverse DNS lookup; otherwise, IP address or
@@ -320,6 +435,113 @@ static constexpr const char *kClientAddress = "client.address";
  proxies, if it's available.</li> </ul>
  */
 static constexpr const char *kClientPort = "client.port";
+
+/**
+ * The cloud account ID the resource is assigned to.
+ */
+static constexpr const char *kCloudAccountId = "cloud.account.id";
+
+/**
+ * Cloud regions often have multiple, isolated locations known as zones to increase availability.
+ Availability zone represents the zone where the resource is running.
+ *
+ * <p>Notes:
+  <ul> <li>Availability zones are called &quot;zones&quot; on Alibaba Cloud and Google Cloud.</li>
+ </ul>
+ */
+static constexpr const char *kCloudAvailabilityZone = "cloud.availability_zone";
+
+/**
+ * The cloud platform in use.
+ *
+ * <p>Notes:
+  <ul> <li>The prefix of the service SHOULD match the one specified in {@code cloud.provider}.</li>
+ </ul>
+ */
+static constexpr const char *kCloudPlatform = "cloud.platform";
+
+/**
+ * Name of the cloud provider.
+ */
+static constexpr const char *kCloudProvider = "cloud.provider";
+
+/**
+ * The geographical region the resource is running.
+ *
+ * <p>Notes:
+  <ul> <li>Refer to your provider's docs to see the available regions, for example <a
+ href="https://www.alibabacloud.com/help/doc-detail/40654.htm">Alibaba Cloud regions</a>, <a
+ href="https://aws.amazon.com/about-aws/global-infrastructure/regions_az/">AWS regions</a>, <a
+ href="https://azure.microsoft.com/global-infrastructure/geographies/">Azure regions</a>, <a
+ href="https://cloud.google.com/about/locations">Google Cloud regions</a>, or <a
+ href="https://www.tencentcloud.com/document/product/213/6091">Tencent Cloud regions</a>.</li> </ul>
+ */
+static constexpr const char *kCloudRegion = "cloud.region";
+
+/**
+ * Cloud provider-specific native identifier of the monitored cloud resource (e.g. an <a
+href="https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html">ARN</a> on AWS, a
+<a href="https://learn.microsoft.com/rest/api/resources/resources/get-by-id">fully qualified
+resource ID</a> on Azure, a <a
+href="https://cloud.google.com/apis/design/resource_names#full_resource_name">full resource name</a>
+on GCP)
+ *
+ * <p>Notes:
+  <ul> <li>On some cloud providers, it may not be possible to determine the full ID at startup,
+so it may be necessary to set {@code cloud.resource_id} as a span attribute instead.</li><li>The
+exact value to use for {@code cloud.resource_id} depends on the cloud provider. The following
+well-known definitions MUST be used if you set this attribute and they apply:</li><li><strong>AWS
+Lambda:</strong> The function <a
+href="https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html">ARN</a>. Take care
+not to use the &quot;invoked ARN&quot; directly but replace any <a
+href="https://docs.aws.amazon.com/lambda/latest/dg/configuration-aliases.html">alias suffix</a> with
+the resolved function version, as the same runtime instance may be invokable with multiple different
+aliases.</li> <li><strong>GCP:</strong> The <a
+href="https://cloud.google.com/iam/docs/full-resource-names">URI of the resource</a></li>
+<li><strong>Azure:</strong> The <a
+href="https://docs.microsoft.com/rest/api/resources/resources/get-by-id">Fully Qualified Resource
+ID</a> of the invoked function, <em>not</em> the function app, having the form
+{@code
+/subscriptions/<SUBSCIPTION_GUID>/resourceGroups/<RG>/providers/Microsoft.Web/sites/<FUNCAPP>/functions/<FUNC>}.
+This means that a span attribute MUST be used, as an Azure function app can host multiple functions
+that would usually share a TracerProvider.</li>
+ </ul>
+ */
+static constexpr const char *kCloudResourceId = "cloud.resource_id";
+
+/**
+ * The <a href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#id">event_id</a>
+ * uniquely identifies the event.
+ */
+static constexpr const char *kCloudeventsEventId = "cloudevents.event_id";
+
+/**
+ * The <a
+ * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1">source</a>
+ * identifies the context in which an event happened.
+ */
+static constexpr const char *kCloudeventsEventSource = "cloudevents.event_source";
+
+/**
+ * The <a
+ * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#specversion">version of
+ * the CloudEvents specification</a> which the event uses.
+ */
+static constexpr const char *kCloudeventsEventSpecVersion = "cloudevents.event_spec_version";
+
+/**
+ * The <a
+ * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#subject">subject</a> of
+ * the event in the context of the event producer (identified by source).
+ */
+static constexpr const char *kCloudeventsEventSubject = "cloudevents.event_subject";
+
+/**
+ * The <a
+ * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#type">event_type</a>
+ * contains a value describing the type of event related to the originating occurrence.
+ */
+static constexpr const char *kCloudeventsEventType = "cloudevents.event_type";
 
 /**
  * The column number in {@code code.filepath} best representing the operation. It SHOULD point
@@ -356,6 +578,87 @@ static constexpr const char *kCodeNamespace = "code.namespace";
  * representation is to be determined and documented by each language SIG.
  */
 static constexpr const char *kCodeStacktrace = "code.stacktrace";
+
+/**
+ * The command used to run the container (i.e. the command name).
+ *
+ * <p>Notes:
+  <ul> <li>If using embedded credentials or sensitive data, it is recommended to remove them to
+ prevent potential leakage.</li> </ul>
+ */
+static constexpr const char *kContainerCommand = "container.command";
+
+/**
+ * All the command arguments (including the command/executable itself) run by the container. [2]
+ */
+static constexpr const char *kContainerCommandArgs = "container.command_args";
+
+/**
+ * The full command run by the container as a single string representing the full command. [2]
+ */
+static constexpr const char *kContainerCommandLine = "container.command_line";
+
+/**
+ * The CPU state for this data point.
+ */
+static constexpr const char *kContainerCpuState = "container.cpu.state";
+
+/**
+ * Container ID. Usually a UUID, as for example used to <a
+ * href="https://docs.docker.com/engine/reference/run/#container-identification">identify Docker
+ * containers</a>. The UUID might be abbreviated.
+ */
+static constexpr const char *kContainerId = "container.id";
+
+/**
+ * Runtime specific image identifier. Usually a hash algorithm followed by a UUID.
+ *
+ * <p>Notes:
+  <ul> <li>Docker defines a sha256 of the image id; {@code container.image.id} corresponds to the
+{@code Image} field from the Docker container inspect <a
+href="https://docs.docker.com/engine/api/v1.43/#tag/Container/operation/ContainerInspect">API</a>
+endpoint. K8s defines a link to the container registry repository with digest {@code "imageID":
+"registry.azurecr.io
+/namespace/service/dockerfile@sha256:bdeabd40c3a8a492eaf9e8e44d0ebbb84bac7ee25ac0cf8a7159d25f62555625"}.
+The ID is assinged by the container runtime and can vary in different environments. Consider using
+{@code oci.manifest.digest} if it is important to identify the same image in different
+environments/runtimes.</li> </ul>
+ */
+static constexpr const char *kContainerImageId = "container.image.id";
+
+/**
+ * Name of the image the container was built on.
+ */
+static constexpr const char *kContainerImageName = "container.image.name";
+
+/**
+ * Repo digests of the container image as provided by the container runtime.
+ *
+ * <p>Notes:
+  <ul> <li><a
+ href="https://docs.docker.com/engine/api/v1.43/#tag/Image/operation/ImageInspect">Docker</a> and <a
+ href="https://github.com/kubernetes/cri-api/blob/c75ef5b473bbe2d0a4fc92f82235efd665ea8e9f/pkg/apis/runtime/v1/api.proto#L1237-L1238">CRI</a>
+ report those under the {@code RepoDigests} field.</li> </ul>
+ */
+static constexpr const char *kContainerImageRepoDigests = "container.image.repo_digests";
+
+/**
+ * Container image tags. An example can be found in <a
+ * href="https://docs.docker.com/engine/api/v1.43/#tag/Image/operation/ImageInspect">Docker Image
+ * Inspect</a>. Should be only the {@code <tag>} section of the full name for example from {@code
+ * registry.example.com/my-org/my-image:<tag>}.
+ */
+static constexpr const char *kContainerImageTags = "container.image.tags";
+
+/**
+ * Container name used by container runtime.
+ */
+static constexpr const char *kContainerName = "container.name";
+
+/**
+ * The container runtime managing this container.
+ */
+static constexpr const char *kContainerRuntime = "container.runtime";
 
 /**
  * The consistency level of the query. Based on consistency values from <a
@@ -404,12 +707,6 @@ static constexpr const char *kDbCassandraSpeculativeExecutionCount =
 static constexpr const char *kDbCassandraTable = "db.cassandra.table";
 
 /**
- * The connection string used to connect to the database. It is recommended to remove embedded
- * credentials.
- */
-static constexpr const char *kDbConnectionString = "db.connection_string";
-
-/**
  * Unique Cosmos client instance id.
  */
 static constexpr const char *kDbCosmosdbClientId = "db.cosmosdb.client_id";
@@ -455,11 +752,6 @@ static constexpr const char *kDbCosmosdbSubStatusCode = "db.cosmosdb.sub_status_
 static constexpr const char *kDbElasticsearchClusterName = "db.elasticsearch.cluster.name";
 
 /**
- * Represents the human-readable identifier of the node/instance to which a request was routed.
- */
-static constexpr const char *kDbElasticsearchNodeName = "db.elasticsearch.node.name";
-
-/**
  * An identifier (address, unique name, or any other identifier) of the database instance that is
  * executing queries or mutations on the current connection. This is useful in cases where the
  * database is running in a clustered environment and the instrumentation is able to record the node
@@ -467,13 +759,6 @@ static constexpr const char *kDbElasticsearchNodeName = "db.elasticsearch.node.n
  * {@code select @@hostname}.
  */
 static constexpr const char *kDbInstanceId = "db.instance.id";
-
-/**
- * The fully-qualified class name of the <a
- * href="https://docs.oracle.com/javase/8/docs/technotes/guides/jdbc/">Java Database Connectivity
- * (JDBC)</a> driver used to connect.
- */
-static constexpr const char *kDbJdbcDriverClassname = "db.jdbc.driver_classname";
 
 /**
  * The MongoDB collection being accessed within the database stated in {@code db.name}.
@@ -552,6 +837,45 @@ static constexpr const char *kDbSystem = "db.system";
 static constexpr const char *kDbUser = "db.user";
 
 /**
+ * Name of the <a href="https://wikipedia.org/wiki/Deployment_environment">deployment
+environment</a> (aka deployment tier).
+ *
+ * <p>Notes:
+  <ul> <li>{@code deployment.environment} does not affect the uniqueness constraints defined through
+the {@code service.namespace}, {@code service.name} and {@code service.instance.id} resource
+attributes. This implies that resources carrying the following attribute combinations MUST be
+considered to be identifying the same service:</li><li>{@code service.name=frontend}, {@code
+deployment.environment=production}</li> <li>{@code service.name=frontend}, {@code
+deployment.environment=staging}.</li>
+ </ul>
+ */
+static constexpr const char *kDeploymentEnvironment = "deployment.environment";
+
+/**
+ * Deprecated, use {@code server.address}, {@code server.port} attributes instead.
+ *
+ * @deprecated Deprecated, use `server.address`, `server.port` attributes instead.
+ */
+OPENTELEMETRY_DEPRECATED
+static constexpr const char *kDbConnectionString = "db.connection_string";
+
+/**
+ * Deprecated, use {@code db.instance.id} instead.
+ *
+ * @deprecated Deprecated, use `db.instance.id` instead.
+ */
+OPENTELEMETRY_DEPRECATED
+static constexpr const char *kDbElasticsearchNodeName = "db.elasticsearch.node.name";
+
+/**
+ * Removed, no replacement at this time.
+ *
+ * @deprecated Removed, no replacement at this time.
+ */
+OPENTELEMETRY_DEPRECATED
+static constexpr const char *kDbJdbcDriverClassname = "db.jdbc.driver_classname";
+
+/**
  * Deprecated, use {@code network.protocol.name} instead.
  *
  * @deprecated Deprecated, use `network.protocol.name` instead.
@@ -622,6 +946,15 @@ static constexpr const char *kHttpUrl = "http.url";
  */
 OPENTELEMETRY_DEPRECATED
 static constexpr const char *kHttpUserAgent = "http.user_agent";
+
+/**
+ * &quot;Deprecated, use {@code messaging.destination.partition.id} instead.&quot;
+ *
+ * @deprecated "Deprecated, use `messaging.destination.partition.id` instead.".
+ */
+OPENTELEMETRY_DEPRECATED
+static constexpr const char *kMessagingKafkaDestinationPartition =
+    "messaging.kafka.destination.partition";
 
 /**
  * Deprecated, use {@code server.address}.
@@ -730,6 +1063,14 @@ OPENTELEMETRY_DEPRECATED
 static constexpr const char *kNetTransport = "net.transport";
 
 /**
+ * Deprecated, use {@code system.process.status} instead.
+ *
+ * @deprecated Deprecated, use `system.process.status` instead.
+ */
+OPENTELEMETRY_DEPRECATED
+static constexpr const char *kSystemProcessesStatus = "system.processes.status";
+
+/**
  * Destination address - domain name if available without reverse DNS lookup; otherwise, IP address
  or Unix domain socket name.
  *
@@ -746,9 +1087,89 @@ static constexpr const char *kDestinationAddress = "destination.address";
 static constexpr const char *kDestinationPort = "destination.port";
 
 /**
+ * A unique identifier representing the device
+ *
+ * <p>Notes:
+  <ul> <li>The device identifier MUST only be defined using the values outlined below. This value is
+ not an advertising identifier and MUST NOT be used as such. On iOS (Swift or Objective-C), this
+ value MUST be equal to the <a
+ href="https://developer.apple.com/documentation/uikit/uidevice/1620059-identifierforvendor">vendor
+ identifier</a>. On Android (Java or Kotlin), this value MUST be equal to the Firebase Installation
+ ID or a globally unique UUID which is persisted across sessions in your application. More
+ information can be found <a
+ href="https://developer.android.com/training/articles/user-data-ids">here</a> on best practices and
+ exact implementation details. Caution should be taken when storing personal data or anything which
+ can identify a user. GDPR and data protection laws may apply, ensure you do your own due
+ diligence.</li> </ul>
+ */
+static constexpr const char *kDeviceId = "device.id";
+
+/**
+ * The name of the device manufacturer
+ *
+ * <p>Notes:
+  <ul> <li>The Android OS provides this field via <a
+ href="https://developer.android.com/reference/android/os/Build#MANUFACTURER">Build</a>. iOS apps
+ SHOULD hardcode the value {@code Apple}.</li> </ul>
+ */
+static constexpr const char *kDeviceManufacturer = "device.manufacturer";
+
+/**
+ * The model identifier for the device
+ *
+ * <p>Notes:
+  <ul> <li>It's recommended this value represents a machine-readable version of the model identifier
+ rather than the market or consumer-friendly name of the device.</li> </ul>
+ */
+static constexpr const char *kDeviceModelIdentifier = "device.model.identifier";
+
+/**
+ * The marketing name for the device model
+ *
+ * <p>Notes:
+  <ul> <li>It's recommended this value represents a human-readable version of the device model
+ rather than a machine-readable alternative.</li> </ul>
+ */
+static constexpr const char *kDeviceModelName = "device.model.name";
+
+/**
  * The disk IO operation direction.
  */
 static constexpr const char *kDiskIoDirection = "disk.io.direction";
+
+/**
+ * The name being queried.
+ *
+ * <p>Notes:
+  <ul> <li>If the name field contains non-printable characters (below 32 or above 126), those
+ characters should be represented as escaped base 10 integers (\DDD). Back slashes and quotes should
+ be escaped. Tabs, carriage returns, and line feeds should be converted to \t, \r, and \n
+ respectively.</li> </ul>
+ */
+static constexpr const char *kDnsQuestionName = "dns.question.name";
+
+/**
+ * Username or client_id extracted from the access token or <a
+ * href="https://tools.ietf.org/html/rfc7235#section-4.2">Authorization</a> header in the inbound
+ * request from outside the system.
+ */
+static constexpr const char *kEnduserId = "enduser.id";
+
+/**
+ * Actual/assumed role the client is making the request under extracted from token or application
+ * security context.
+ */
+static constexpr const char *kEnduserRole = "enduser.role";
+
+/**
+ * Scopes or granted authorities the client currently possesses extracted from token or application
+ * security context. The value would come from the scope associated with an <a
+ * href="https://tools.ietf.org/html/rfc6749#section-3.3">OAuth 2.0 Access Token</a> or an attribute
+ * value in a <a
+ * href="http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-tech-overview-2.0.html">SAML 2.0
+ * Assertion</a>.
+ */
+static constexpr const char *kEnduserScope = "enduser.scope";
 
 /**
  * Describes a class of error the operation ended with.
@@ -806,6 +1227,331 @@ static constexpr const char *kExceptionStacktrace = "exception.stacktrace";
 static constexpr const char *kExceptionType = "exception.type";
 
 /**
+ * A boolean that is true if the serverless function is executed for the first time (aka
+ * cold-start).
+ */
+static constexpr const char *kFaasColdstart = "faas.coldstart";
+
+/**
+ * A string containing the schedule period as <a
+ * href="https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm">Cron
+ * Expression</a>.
+ */
+static constexpr const char *kFaasCron = "faas.cron";
+
+/**
+ * The name of the source on which the triggering operation was performed. For example, in Cloud
+ * Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.
+ */
+static constexpr const char *kFaasDocumentCollection = "faas.document.collection";
+
+/**
+ * The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the
+ * name of the file, and in Cosmos DB the table name.
+ */
+static constexpr const char *kFaasDocumentName = "faas.document.name";
+
+/**
+ * Describes the type of the operation that was performed on the data.
+ */
+static constexpr const char *kFaasDocumentOperation = "faas.document.operation";
+
+/**
+ * A string containing the time when the data was accessed in the <a
+ * href="https://www.iso.org/iso-8601-date-and-time-format.html">ISO 8601</a> format expressed in <a
+ * href="https://www.w3.org/TR/NOTE-datetime">UTC</a>.
+ */
+static constexpr const char *kFaasDocumentTime = "faas.document.time";
+
+/**
+ * The execution environment ID as a string, that will be potentially reused for other invocations
+ to the same function/function version.
+ *
+ * <p>Notes:
+  <ul> <li><strong>AWS Lambda:</strong> Use the (full) log stream name.</li>
+ </ul>
+ */
+static constexpr const char *kFaasInstance = "faas.instance";
+
+/**
+ * The invocation ID of the current function invocation.
+ */
+static constexpr const char *kFaasInvocationId = "faas.invocation_id";
+
+/**
+ * The name of the invoked function.
+ *
+ * <p>Notes:
+  <ul> <li>SHOULD be equal to the {@code faas.name} resource attribute of the invoked function.</li>
+ </ul>
+ */
+static constexpr const char *kFaasInvokedName = "faas.invoked_name";
+
+/**
+ * The cloud provider of the invoked function.
+ *
+ * <p>Notes:
+  <ul> <li>SHOULD be equal to the {@code cloud.provider} resource attribute of the invoked
+ function.</li> </ul>
+ */
+static constexpr const char *kFaasInvokedProvider = "faas.invoked_provider";
+
+/**
+ * The cloud region of the invoked function.
+ *
+ * <p>Notes:
+  <ul> <li>SHOULD be equal to the {@code cloud.region} resource attribute of the invoked
+ function.</li> </ul>
+ */
+static constexpr const char *kFaasInvokedRegion = "faas.invoked_region";
+
+/**
+ * The amount of memory available to the serverless function converted to Bytes.
+ *
+ * <p>Notes:
+  <ul> <li>It's recommended to set this attribute since e.g. too little memory can easily stop a
+ Java AWS Lambda function from working correctly. On AWS Lambda, the environment variable {@code
+ AWS_LAMBDA_FUNCTION_MEMORY_SIZE} provides this information (which must be multiplied by
+ 1,048,576).</li> </ul>
+ */
+static constexpr const char *kFaasMaxMemory = "faas.max_memory";
+
+/**
+ * The name of the single function that this runtime instance executes.
+ *
+ * <p>Notes:
+  <ul> <li>This is the name of the function as configured/deployed on the FaaS
+platform and is usually different from the name of the callback
+function (which may be stored in the
+<a href="/docs/general/attributes.md#source-code-attributes">{@code code.namespace}/{@code
+code.function}</a> span attributes).</li><li>For some cloud providers, the above definition is
+ambiguous. The following definition of function name MUST be used for this attribute (and
+consequently the span name) for the listed cloud providers/products:</li><li><strong>Azure:</strong>
+The full name {@code <FUNCAPP>/<FUNC>}, i.e., function app name followed by a forward slash followed
+by the function name (this form can also be seen in the resource JSON for the function). This means
+that a span attribute MUST be used, as an Azure function app can host multiple functions that would
+usually share a TracerProvider (see also the {@code cloud.resource_id} attribute).</li>
+ </ul>
+ */
+static constexpr const char *kFaasName = "faas.name";
+
+/**
+ * A string containing the function invocation time in the <a
+ * href="https://www.iso.org/iso-8601-date-and-time-format.html">ISO 8601</a> format expressed in <a
+ * href="https://www.w3.org/TR/NOTE-datetime">UTC</a>.
+ */
+static constexpr const char *kFaasTime = "faas.time";
+
+/**
+ * Type of the trigger which caused this function invocation.
+ */
+static constexpr const char *kFaasTrigger = "faas.trigger";
+
+/**
+ * The immutable version of the function being executed.
+ *
+ * <p>Notes:
+  <ul> <li>Depending on the cloud provider and platform, use:</li><li><strong>AWS Lambda:</strong>
+The <a href="https://docs.aws.amazon.com/lambda/latest/dg/configuration-versions.html">function
+version</a> (an integer represented as a decimal string).</li> <li><strong>Google Cloud Run
+(Services):</strong> The <a href="https://cloud.google.com/run/docs/managing/revisions">revision</a>
+(i.e., the function name plus the revision suffix).</li>
+<li><strong>Google Cloud Functions:</strong> The value of the
+<a
+href="https://cloud.google.com/functions/docs/env-var#runtime_environment_variables_set_automatically">{@code
+K_REVISION} environment variable</a>.</li> <li><strong>Azure Functions:</strong> Not applicable. Do
+not set this attribute.</li>
+ </ul>
+ */
+static constexpr const char *kFaasVersion = "faas.version";
+
+/**
+ * The unique identifier of the feature flag.
+ */
+static constexpr const char *kFeatureFlagKey = "feature_flag.key";
+
+/**
+ * The name of the service provider that performs the flag evaluation.
+ */
+static constexpr const char *kFeatureFlagProviderName = "feature_flag.provider_name";
+
+/**
+ * SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the
+value can be used.
+ *
+ * <p>Notes:
+  <ul> <li>A semantic identifier, commonly referred to as a variant, provides a means
+for referring to a value without including the value itself. This can
+provide additional context for understanding the meaning behind a value.
+For example, the variant {@code red} maybe be used for the value {@code #c05543}.</li><li>A
+stringified version of the value can be used in situations where a semantic identifier is
+unavailable. String representation of the value should be determined by the implementer.</li> </ul>
+ */
+static constexpr const char *kFeatureFlagVariant = "feature_flag.variant";
+
+/**
+ * Directory where the file is located. It should include the drive letter, when appropriate.
+ */
+static constexpr const char *kFileDirectory = "file.directory";
+
+/**
+ * File extension, excluding the leading dot.
+ *
+ * <p>Notes:
+  <ul> <li>When the file name has multiple extensions (example.tar.gz), only the last one should be
+ captured (&quot;gz&quot;, not &quot;tar.gz&quot;).</li> </ul>
+ */
+static constexpr const char *kFileExtension = "file.extension";
+
+/**
+ * Name of the file including the extension, without the directory.
+ */
+static constexpr const char *kFileName = "file.name";
+
+/**
+ * Full path to the file, including the file name. It should include the drive letter, when
+ * appropriate.
+ */
+static constexpr const char *kFilePath = "file.path";
+
+/**
+ * File size in bytes.
+ */
+static constexpr const char *kFileSize = "file.size";
+
+/**
+ * The name of the Cloud Run <a
+ * href="https://cloud.google.com/run/docs/managing/job-executions">execution</a> being run for the
+ * Job, as set by the <a
+ * href="https://cloud.google.com/run/docs/container-contract#jobs-env-vars">{@code
+ * CLOUD_RUN_EXECUTION}</a> environment variable.
+ */
+static constexpr const char *kGcpCloudRunJobExecution = "gcp.cloud_run.job.execution";
+
+/**
+ * The index for a task within an execution as provided by the <a
+ * href="https://cloud.google.com/run/docs/container-contract#jobs-env-vars">{@code
+ * CLOUD_RUN_TASK_INDEX}</a> environment variable.
+ */
+static constexpr const char *kGcpCloudRunJobTaskIndex = "gcp.cloud_run.job.task_index";
+
+/**
+ * The hostname of a GCE instance. This is the full value of the default or <a
+ * href="https://cloud.google.com/compute/docs/instances/custom-hostname-vm">custom hostname</a>.
+ */
+static constexpr const char *kGcpGceInstanceHostname = "gcp.gce.instance.hostname";
+
+/**
+ * The instance name of a GCE instance. This is the value provided by {@code host.name}, the visible
+ * name of the instance in the Cloud Console UI, and the prefix for the default hostname of the
+ * instance as defined by the <a
+ * href="https://cloud.google.com/compute/docs/internal-dns#instance-fully-qualified-domain-names">default
+ * internal DNS name</a>.
+ */
+static constexpr const char *kGcpGceInstanceName = "gcp.gce.instance.name";
+
+/**
+ * The CPU architecture the host system is running on.
+ */
+static constexpr const char *kHostArch = "host.arch";
+
+/**
+ * The amount of level 2 memory cache available to the processor (in Bytes).
+ */
+static constexpr const char *kHostCpuCacheL2Size = "host.cpu.cache.l2.size";
+
+/**
+ * Family or generation of the CPU.
+ */
+static constexpr const char *kHostCpuFamily = "host.cpu.family";
+
+/**
+ * Model identifier. It provides more granular information about the CPU, distinguishing it from
+ * other CPUs within the same family.
+ */
+static constexpr const char *kHostCpuModelId = "host.cpu.model.id";
+
+/**
+ * Model designation of the processor.
+ */
+static constexpr const char *kHostCpuModelName = "host.cpu.model.name";
+
+/**
+ * Stepping or core revisions.
+ */
+static constexpr const char *kHostCpuStepping = "host.cpu.stepping";
+
+/**
+ * Processor manufacturer identifier. A maximum 12-character string.
+ *
+ * <p>Notes:
+  <ul> <li><a href="https://wiki.osdev.org/CPUID">CPUID</a> command returns the vendor ID string in
+ EBX, EDX and ECX registers. Writing these to memory in this order results in a 12-character
+ string.</li> </ul>
+ */
+static constexpr const char *kHostCpuVendorId = "host.cpu.vendor.id";
+
+/**
+ * Unique host ID. For Cloud, this must be the instance_id assigned by the cloud provider. For
+ * non-containerized systems, this should be the {@code machine-id}. See the table below for the
+ * sources to use to determine the {@code machine-id} based on operating system.
+ */
+static constexpr const char *kHostId = "host.id";
+
+/**
+ * VM image ID or host OS image ID. For Cloud, this value is from the provider.
+ */
+static constexpr const char *kHostImageId = "host.image.id";
+
+/**
+ * Name of the VM image or OS install the host was instantiated from.
+ */
+static constexpr const char *kHostImageName = "host.image.name";
+
+/**
+ * The version string of the VM image or host OS as defined in <a
+ * href="/docs/resource/README.md#version-attributes">Version Attributes</a>.
+ */
+static constexpr const char *kHostImageVersion = "host.image.version";
+
+/**
+ * Available IP addresses of the host, excluding loopback interfaces.
+ *
+ * <p>Notes:
+  <ul> <li>IPv4 Addresses MUST be specified in dotted-quad notation. IPv6 addresses MUST be
+ specified in the <a href="https://www.rfc-editor.org/rfc/rfc5952.html">RFC 5952</a> format.</li>
+ </ul>
+ */
+static constexpr const char *kHostIp = "host.ip";
+
+/**
+ * Available MAC addresses of the host, excluding loopback interfaces.
+ *
+ * <p>Notes:
+  <ul> <li>MAC Addresses MUST be represented in <a
+ href="https://standards.ieee.org/wp-content/uploads/import/documents/tutorials/eui.pdf">IEEE RA
+ hexadecimal form</a>: as hyphen-separated octets in uppercase hexadecimal form from most to least
+ significant.</li> </ul>
+ */
+static constexpr const char *kHostMac = "host.mac";
+
+/**
+ * Name of the host. On Unix systems, it may contain what the hostname command returns, or the fully
+ * qualified hostname, or another name specified by the user.
+ */
+static constexpr const char *kHostName = "host.name";
+
+/**
+ * Type of host. For Cloud, this must be the machine type.
+ */
+static constexpr const char *kHostType = "host.type";
+
+/**
+ * State of the HTTP connection in the HTTP connection pool.
+ */
+static constexpr const char *kHttpConnectionState = "http.connection.state";
+
+/**
  * The size of the request payload body in bytes. This is the number of bytes transferred excluding
  * headers and is often, but not always, present as the <a
  * href="https://www.rfc-editor.org/rfc/rfc9110.html#field.content-length">Content-Length</a>
@@ -851,12 +1597,26 @@ static constexpr const char *kHttpRequestMethodOriginal = "http.request.method_o
 static constexpr const char *kHttpRequestResendCount = "http.request.resend_count";
 
 /**
+ * The total size of the request in bytes. This should be the total number of bytes sent over the
+ * wire, including the request line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and request
+ * body if any.
+ */
+static constexpr const char *kHttpRequestSize = "http.request.size";
+
+/**
  * The size of the response payload body in bytes. This is the number of bytes transferred excluding
  * headers and is often, but not always, present as the <a
  * href="https://www.rfc-editor.org/rfc/rfc9110.html#field.content-length">Content-Length</a>
  * header. For requests using transport encoding, this should be the compressed size.
  */
 static constexpr const char *kHttpResponseBodySize = "http.response.body.size";
+
+/**
+ * The total size of the response in bytes. This should be the total number of bytes sent over the
+ * wire, including the status line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and response
+ * body and trailers if any.
+ */
+static constexpr const char *kHttpResponseSize = "http.response.size";
 
 /**
  * <a href="https://tools.ietf.org/html/rfc7231#section-6">HTTP response status code</a>.
@@ -874,6 +1634,132 @@ the <a href="/docs/http/http-spans.md#http-server-definitions">application root<
 one.</li> </ul>
  */
 static constexpr const char *kHttpRoute = "http.route";
+
+/**
+ * The name of the cluster.
+ */
+static constexpr const char *kK8sClusterName = "k8s.cluster.name";
+
+/**
+ * A pseudo-ID for the cluster, set to the UID of the {@code kube-system} namespace.
+ *
+ * <p>Notes:
+  <ul> <li>K8s doesn't have support for obtaining a cluster ID. If this is ever
+added, we will recommend collecting the {@code k8s.cluster.uid} through the
+official APIs. In the meantime, we are able to use the {@code uid} of the
+{@code kube-system} namespace as a proxy for cluster ID. Read on for the
+rationale.</li><li>Every object created in a K8s cluster is assigned a distinct UID. The
+{@code kube-system} namespace is used by Kubernetes itself and will exist
+for the lifetime of the cluster. Using the {@code uid} of the {@code kube-system}
+namespace is a reasonable proxy for the K8s ClusterID as it will only
+change if the cluster is rebuilt. Furthermore, Kubernetes UIDs are
+UUIDs as standardized by
+<a href="https://www.itu.int/ITU-T/studygroups/com17/oid.html">ISO/IEC 9834-8 and ITU-T X.667</a>.
+Which states:</li><blockquote>
+<li>If generated according to one of the mechanisms defined in Rec.</li></blockquote>
+<li>ITU-T X.667 | ISO/IEC 9834-8, a UUID is either guaranteed to be
+  different from all other UUIDs generated before 3603 A.D., or is
+  extremely likely to be different (depending on the mechanism chosen).</li><li>Therefore, UIDs
+between clusters should be extremely unlikely to conflict.</li> </ul>
+ */
+static constexpr const char *kK8sClusterUid = "k8s.cluster.uid";
+
+/**
+ * The name of the Container from Pod specification, must be unique within a Pod. Container runtime
+ * usually uses different globally unique name ({@code container.name}).
+ */
+static constexpr const char *kK8sContainerName = "k8s.container.name";
+
+/**
+ * Number of times the container was restarted. This attribute can be used to identify a particular
+ * container (running or stopped) within a container spec.
+ */
+static constexpr const char *kK8sContainerRestartCount = "k8s.container.restart_count";
+
+/**
+ * The name of the CronJob.
+ */
+static constexpr const char *kK8sCronjobName = "k8s.cronjob.name";
+
+/**
+ * The UID of the CronJob.
+ */
+static constexpr const char *kK8sCronjobUid = "k8s.cronjob.uid";
+
+/**
+ * The name of the DaemonSet.
+ */
+static constexpr const char *kK8sDaemonsetName = "k8s.daemonset.name";
+
+/**
+ * The UID of the DaemonSet.
+ */
+static constexpr const char *kK8sDaemonsetUid = "k8s.daemonset.uid";
+
+/**
+ * The name of the Deployment.
+ */
+static constexpr const char *kK8sDeploymentName = "k8s.deployment.name";
+
+/**
+ * The UID of the Deployment.
+ */
+static constexpr const char *kK8sDeploymentUid = "k8s.deployment.uid";
+
+/**
+ * The name of the Job.
+ */
+static constexpr const char *kK8sJobName = "k8s.job.name";
+
+/**
+ * The UID of the Job.
+ */
+static constexpr const char *kK8sJobUid = "k8s.job.uid";
+
+/**
+ * The name of the namespace that the pod is running in.
+ */
+static constexpr const char *kK8sNamespaceName = "k8s.namespace.name";
+
+/**
+ * The name of the Node.
+ */
+static constexpr const char *kK8sNodeName = "k8s.node.name";
+
+/**
+ * The UID of the Node.
+ */
+static constexpr const char *kK8sNodeUid = "k8s.node.uid";
+
+/**
+ * The name of the Pod.
+ */
+static constexpr const char *kK8sPodName = "k8s.pod.name";
+
+/**
+ * The UID of the Pod.
+ */
+static constexpr const char *kK8sPodUid = "k8s.pod.uid";
+
+/**
+ * The name of the ReplicaSet.
+ */
+static constexpr const char *kK8sReplicasetName = "k8s.replicaset.name";
+
+/**
+ * The UID of the ReplicaSet.
+ */
+static constexpr const char *kK8sReplicasetUid = "k8s.replicaset.uid";
+
+/**
+ * The name of the StatefulSet.
+ */
+static constexpr const char *kK8sStatefulsetName = "k8s.statefulset.name";
+
+/**
+ * The UID of the StatefulSet.
+ */
+static constexpr const char *kK8sStatefulsetUid = "k8s.statefulset.uid";
 
 /**
  * The number of messages sent, received, or processed in the scope of the batching operation.
@@ -907,6 +1793,13 @@ the broker. If the broker doesn't have such notion, the destination name SHOULD 
 the broker.</li> </ul>
  */
 static constexpr const char *kMessagingDestinationName = "messaging.destination.name";
+
+/**
+ * The identifier of the partition messages are sent to or received from, unique within the {@code
+ * messaging.destination.name}.
+ */
+static constexpr const char *kMessagingDestinationPartitionId =
+    "messaging.destination.partition.id";
 
 /**
  * Low cardinality representation of the messaging destination name
@@ -944,6 +1837,18 @@ static constexpr const char *kMessagingDestinationPublishName =
     "messaging.destination_publish.name";
 
 /**
+ * The name of the consumer group the event consumer is associated with.
+ */
+static constexpr const char *kMessagingEventhubsConsumerGroup =
+    "messaging.eventhubs.consumer.group";
+
+/**
+ * The UTC epoch seconds at which the message has been accepted and stored in the entity.
+ */
+static constexpr const char *kMessagingEventhubsMessageEnqueuedTime =
+    "messaging.eventhubs.message.enqueued_time";
+
+/**
  * The ordering key for a given message. If the attribute is not present, the message does not have
  * an ordering key.
  */
@@ -955,12 +1860,6 @@ static constexpr const char *kMessagingGcpPubsubMessageOrderingKey =
  * producers.
  */
 static constexpr const char *kMessagingKafkaConsumerGroup = "messaging.kafka.consumer.group";
-
-/**
- * Partition the message is sent to.
- */
-static constexpr const char *kMessagingKafkaDestinationPartition =
-    "messaging.kafka.destination.partition";
 
 /**
  * Message keys in Kafka are used for grouping alike messages to ensure they're processed on the
@@ -1028,6 +1927,12 @@ static constexpr const char *kMessagingRabbitmqDestinationRoutingKey =
     "messaging.rabbitmq.destination.routing_key";
 
 /**
+ * RabbitMQ message delivery tag
+ */
+static constexpr const char *kMessagingRabbitmqMessageDeliveryTag =
+    "messaging.rabbitmq.message.delivery_tag";
+
+/**
  * Name of the RocketMQ producer/consumer group that is handling the message. The client type is
  * identified by the SpanKind.
  */
@@ -1076,6 +1981,32 @@ static constexpr const char *kMessagingRocketmqMessageType = "messaging.rocketmq
  * Namespace of RocketMQ resources, resources in different namespaces are individual.
  */
 static constexpr const char *kMessagingRocketmqNamespace = "messaging.rocketmq.namespace";
+
+/**
+ * The name of the subscription in the topic messages are received from.
+ */
+static constexpr const char *kMessagingServicebusDestinationSubscriptionName =
+    "messaging.servicebus.destination.subscription_name";
+
+/**
+ * Describes the <a
+ * href="https://learn.microsoft.com/azure/service-bus-messaging/message-transfers-locks-settlement#peeklock">settlement
+ * type</a>.
+ */
+static constexpr const char *kMessagingServicebusDispositionStatus =
+    "messaging.servicebus.disposition_status";
+
+/**
+ * Number of deliveries that have been attempted for this message.
+ */
+static constexpr const char *kMessagingServicebusMessageDeliveryCount =
+    "messaging.servicebus.message.delivery_count";
+
+/**
+ * The UTC epoch seconds at which the message has been accepted and stored in the entity.
+ */
+static constexpr const char *kMessagingServicebusMessageEnqueuedTime =
+    "messaging.servicebus.message.enqueued_time";
 
 /**
  * An identifier for the messaging system being used. See below for a list of well-known
@@ -1149,12 +2080,13 @@ static constexpr const char *kNetworkPeerPort = "network.peer.port";
 static constexpr const char *kNetworkProtocolName = "network.protocol.name";
 
 /**
- * Version of the protocol specified in {@code network.protocol.name}.
+ * The actual version of the protocol used for network communication.
  *
  * <p>Notes:
-  <ul> <li>{@code network.protocol.version} refers to the version of the protocol used and might be
- different from the protocol client's version. If the HTTP client has a version of {@code 0.27.2},
- but sends HTTP version {@code 1.1}, this attribute should be set to {@code 1.1}.</li> </ul>
+  <ul> <li>If protocol version is subject to negotiation (for example using <a
+ href="https://www.rfc-editor.org/rfc/rfc7301.html">ALPN</a>), this attribute SHOULD be set to the
+ negotiated version. If the actual protocol version is not known, this attribute SHOULD NOT be
+ set.</li> </ul>
  */
 static constexpr const char *kNetworkProtocolVersion = "network.protocol.version";
 
@@ -1178,6 +2110,121 @@ static constexpr const char *kNetworkTransport = "network.transport";
   <ul> <li>The value SHOULD be normalized to lowercase.</li> </ul>
  */
 static constexpr const char *kNetworkType = "network.type";
+
+/**
+ * The digest of the OCI image manifest. For container images specifically is the digest by which
+the container image is known.
+ *
+ * <p>Notes:
+  <ul> <li>Follows <a href="https://github.com/opencontainers/image-spec/blob/main/manifest.md">OCI
+Image Manifest Specification</a>, and specifically the <a
+href="https://github.com/opencontainers/image-spec/blob/main/descriptor.md#digests">Digest
+property</a>. An example can be found in <a
+href="https://docs.docker.com/registry/spec/manifest-v2-2/#example-image-manifest">Example Image
+Manifest</a>.</li> </ul>
+ */
+static constexpr const char *kOciManifestDigest = "oci.manifest.digest";
+
+/**
+ * Unique identifier for a particular build or compilation of the operating system.
+ */
+static constexpr const char *kOsBuildId = "os.build_id";
+
+/**
+ * Human readable (not intended to be parsed) OS version information, like e.g. reported by {@code
+ * ver} or {@code lsb_release -a} commands.
+ */
+static constexpr const char *kOsDescription = "os.description";
+
+/**
+ * Human readable operating system name.
+ */
+static constexpr const char *kOsName = "os.name";
+
+/**
+ * The operating system type.
+ */
+static constexpr const char *kOsType = "os.type";
+
+/**
+ * The version string of the operating system as defined in <a
+ * href="/docs/resource/README.md#version-attributes">Version Attributes</a>.
+ */
+static constexpr const char *kOsVersion = "os.version";
+
+/**
+ * The <a href="/docs/resource/README.md#service">{@code service.name}</a> of the remote service.
+ * SHOULD be equal to the actual {@code service.name} resource attribute of the remote service if
+ * any.
+ */
+static constexpr const char *kPeerService = "peer.service";
+
+/**
+ * The command used to launch the process (i.e. the command name). On Linux based systems, can be
+ * set to the zeroth string in {@code proc/[pid]/cmdline}. On Windows, can be set to the first
+ * parameter extracted from {@code GetCommandLineW}.
+ */
+static constexpr const char *kProcessCommand = "process.command";
+
+/**
+ * All the command arguments (including the command/executable itself) as received by the process.
+ * On Linux-based systems (and some other Unixoid systems supporting procfs), can be set according
+ * to the list of null-delimited strings extracted from {@code proc/[pid]/cmdline}. For libc-based
+ * executables, this would be the full argv vector passed to {@code main}.
+ */
+static constexpr const char *kProcessCommandArgs = "process.command_args";
+
+/**
+ * The full command used to launch the process as a single string representing the full command. On
+ * Windows, can be set to the result of {@code GetCommandLineW}. Do not set this if you have to
+ * assemble it just for monitoring; use {@code process.command_args} instead.
+ */
+static constexpr const char *kProcessCommandLine = "process.command_line";
+
+/**
+ * The name of the process executable. On Linux based systems, can be set to the {@code Name} in
+ * {@code proc/[pid]/status}. On Windows, can be set to the base name of {@code
+ * GetProcessImageFileNameW}.
+ */
+static constexpr const char *kProcessExecutableName = "process.executable.name";
+
+/**
+ * The full path to the process executable. On Linux based systems, can be set to the target of
+ * {@code proc/[pid]/exe}. On Windows, can be set to the result of {@code GetProcessImageFileNameW}.
+ */
+static constexpr const char *kProcessExecutablePath = "process.executable.path";
+
+/**
+ * The username of the user that owns the process.
+ */
+static constexpr const char *kProcessOwner = "process.owner";
+
+/**
+ * Parent Process identifier (PPID).
+ */
+static constexpr const char *kProcessParentPid = "process.parent_pid";
+
+/**
+ * Process identifier (PID).
+ */
+static constexpr const char *kProcessPid = "process.pid";
+
+/**
+ * An additional description about the runtime of the process, for example a specific vendor
+ * customization of the runtime environment.
+ */
+static constexpr const char *kProcessRuntimeDescription = "process.runtime.description";
+
+/**
+ * The name of the runtime of this process. For compiled native binaries, this SHOULD be the name of
+ * the compiler.
+ */
+static constexpr const char *kProcessRuntimeName = "process.runtime.name";
+
+/**
+ * The version of the runtime of this process, as returned by the runtime without modification.
+ */
+static constexpr const char *kProcessRuntimeVersion = "process.runtime.version";
 
 /**
  * The <a href="https://connect.build/docs/protocol/#error-codes">error codes</a> of the Connect
@@ -1265,6 +2312,77 @@ static constexpr const char *kServerAddress = "server.address";
 static constexpr const char *kServerPort = "server.port";
 
 /**
+ * The string ID of the service instance.
+ *
+ * <p>Notes:
+  <ul> <li>MUST be unique for each instance of the same {@code service.namespace,service.name} pair
+(in other words
+{@code service.namespace,service.name,service.instance.id} triplet MUST be globally unique). The ID
+helps to distinguish instances of the same service that exist at the same time (e.g. instances of a
+horizontally scaled service).</li><li>Implementations, such as SDKs, are recommended to generate a
+random Version 1 or Version 4 <a href="https://www.ietf.org/rfc/rfc4122.txt">RFC 4122</a> UUID, but
+are free to use an inherent unique ID as the source of this value if stability is desirable. In that
+case, the ID SHOULD be used as source of a UUID Version 5 and SHOULD use the following UUID as the
+namespace: {@code 4d63009a-8d0f-11ee-aad7-4c796ed8e320}.</li><li>UUIDs are typically recommended, as
+only an opaque value for the purposes of identifying a service instance is needed. Similar to what
+can be seen in the man page for the <a
+href="https://www.freedesktop.org/software/systemd/man/machine-id.html">{@code /etc/machine-id}</a>
+file, the underlying data, such as pod name and namespace should be treated as confidential, being
+the user's choice to expose it or not via another resource attribute.</li><li>For applications
+running behind an application server (like unicorn), we do not recommend using one identifier for
+all processes participating in the application. Instead, it's recommended each division (e.g. a
+worker thread in unicorn) to have its own instance.id.</li><li>It's not recommended for a Collector
+to set {@code service.instance.id} if it can't unambiguously determine the service instance that is
+generating that telemetry. For instance, creating an UUID based on {@code pod.name} will likely be
+wrong, as the Collector might not know from which container within that pod the telemetry
+originated. However, Collectors can set the {@code service.instance.id} if they can unambiguously
+determine the service instance for that telemetry. This is typically the case for scraping
+receivers, as they know the target address and port.</li> </ul>
+ */
+static constexpr const char *kServiceInstanceId = "service.instance.id";
+
+/**
+ * Logical name of the service.
+ *
+ * <p>Notes:
+  <ul> <li>MUST be the same for all instances of horizontally scaled services. If the value was not
+ specified, SDKs MUST fallback to {@code unknown_service:} concatenated with <a
+ href="process.md#process">{@code process.executable.name}</a>, e.g. {@code unknown_service:bash}.
+ If {@code process.executable.name} is not available, the value MUST be set to {@code
+ unknown_service}.</li> </ul>
+ */
+static constexpr const char *kServiceName = "service.name";
+
+/**
+ * A namespace for {@code service.name}.
+ *
+ * <p>Notes:
+  <ul> <li>A string value having a meaning that helps to distinguish a group of services, for
+ example the team name that owns a group of services. {@code service.name} is expected to be unique
+ within the same namespace. If {@code service.namespace} is not specified in the Resource then
+ {@code service.name} is expected to be unique for all services that have no explicit namespace
+ defined (so the empty/unspecified namespace is simply one more valid namespace). Zero-length
+ namespace string is assumed equal to unspecified namespace.</li> </ul>
+ */
+static constexpr const char *kServiceNamespace = "service.namespace";
+
+/**
+ * The version string of the service API or implementation. The format is not defined by these
+ * conventions.
+ */
+static constexpr const char *kServiceVersion = "service.version";
+
+/**
+ * A unique id to identify a session.
+ */
+static constexpr const char *kSessionId = "session.id";
+
+/**
+ * The previous {@code session.id} for this user, when known.
+ */
+static constexpr const char *kSessionPreviousId = "session.previous_id";
+
+/**
  * Source address - domain name if available without reverse DNS lookup; otherwise, IP address or
  Unix domain socket name.
  *
@@ -1279,6 +2397,45 @@ static constexpr const char *kSourceAddress = "source.address";
  * Source port number
  */
 static constexpr const char *kSourcePort = "source.port";
+
+/**
+ * The language of the telemetry SDK.
+ */
+static constexpr const char *kTelemetrySdkLanguage = "telemetry.sdk.language";
+
+/**
+ * The name of the telemetry SDK as defined above.
+ *
+ * <p>Notes:
+  <ul> <li>The OpenTelemetry SDK MUST set the {@code telemetry.sdk.name} attribute to {@code
+opentelemetry}. If another SDK, like a fork or a vendor-provided implementation, is used, this SDK
+MUST set the
+{@code telemetry.sdk.name} attribute to the fully-qualified class or module name of this SDK's main
+entry point or another suitable identifier depending on the language. The identifier {@code
+opentelemetry} is reserved and MUST NOT be used in this case. All custom identifiers SHOULD be
+stable across different versions of an implementation.</li> </ul>
+ */
+static constexpr const char *kTelemetrySdkName = "telemetry.sdk.name";
+
+/**
+ * The version string of the telemetry SDK.
+ */
+static constexpr const char *kTelemetrySdkVersion = "telemetry.sdk.version";
+
+/**
+ * The name of the auto instrumentation agent or distribution, if used.
+ *
+ * <p>Notes:
+  <ul> <li>Official auto instrumentation agents and distributions SHOULD set the {@code
+telemetry.distro.name} attribute to a string starting with {@code opentelemetry-}, e.g. {@code
+opentelemetry-java-instrumentation}.</li> </ul>
+ */
+static constexpr const char *kTelemetryDistroName = "telemetry.distro.name";
+
+/**
+ * The version string of the auto instrumentation agent or distribution, if used.
+ */
+static constexpr const char *kTelemetryDistroVersion = "telemetry.distro.version";
 
 /**
  * Current &quot;managed&quot; thread ID (as opposed to OS thread ID).
@@ -1474,6 +2631,28 @@ static constexpr const char *kTlsServerNotBefore = "tls.server.not_before";
 static constexpr const char *kTlsServerSubject = "tls.server.subject";
 
 /**
+ * Domain extracted from the {@code url.full}, such as &quot;opentelemetry.io&quot;.
+ *
+ * <p>Notes:
+  <ul> <li>In some cases a URL may refer to an IP and/or port directly, without a domain name. In
+ this case, the IP address would go to the domain field. If the URL contains a <a
+ href="https://www.rfc-editor.org/rfc/rfc2732#section-2">literal IPv6 address</a> enclosed by {@code
+ [} and {@code ]}, the {@code [} and {@code ]} characters should also be captured in the domain
+ field.</li> </ul>
+ */
+static constexpr const char *kUrlDomain = "url.domain";
+
+/**
+ * The file extension extracted from the {@code url.full}, excluding the leading dot.
+ *
+ * <p>Notes:
+  <ul> <li>The file extension is only set if it exists, as not every url has a file extension. When
+ the file name has multiple extensions {@code example.tar.gz}, only the last one should be captured
+ {@code gz}, not {@code tar.gz}.</li> </ul>
+ */
+static constexpr const char *kUrlExtension = "url.extension";
+
+/**
  * The <a href="https://www.rfc-editor.org/rfc/rfc3986#section-3.5">URI fragment</a> component
  */
 static constexpr const char *kUrlFragment = "url.fragment";
@@ -1489,24 +2668,58 @@ nevertheless.
 {@code url.full} MUST NOT contain credentials passed via URL in form of {@code
 https://username:password@www.example.com/}. In such case username and password SHOULD be redacted
 and attribute's value SHOULD be {@code https://REDACTED:REDACTED@www.example.com/}.
-{@code url.full} SHOULD capture the absolute URL when it is available (or can be reconstructed) and
-SHOULD NOT be validated or modified except for sanitizing purposes.</li> </ul>
+{@code url.full} SHOULD capture the absolute URL when it is available (or can be reconstructed).
+Sensitive content provided in {@code url.full} SHOULD be scrubbed when instrumentations can identify
+it.</li> </ul>
  */
 static constexpr const char *kUrlFull = "url.full";
 
 /**
+ * Unmodified original URL as seen in the event source.
+ *
+ * <p>Notes:
+  <ul> <li>In network monitoring, the observed URL may be a full URL, whereas in access logs, the
+URL is often just represented as a path. This field is meant to represent the URL as it was
+observed, complete or not.
+{@code url.original} might contain credentials passed via URL in form of {@code
+https://username:password@www.example.com/}. In such case password and username SHOULD NOT be
+redacted and attribute's value SHOULD remain the same.</li> </ul>
+ */
+static constexpr const char *kUrlOriginal = "url.original";
+
+/**
  * The <a href="https://www.rfc-editor.org/rfc/rfc3986#section-3.3">URI path</a> component
+ *
+ * <p>Notes:
+  <ul> <li>Sensitive content provided in {@code url.path} SHOULD be scrubbed when instrumentations
+ can identify it.</li> </ul>
  */
 static constexpr const char *kUrlPath = "url.path";
+
+/**
+ * Port extracted from the {@code url.full}
+ */
+static constexpr const char *kUrlPort = "url.port";
 
 /**
  * The <a href="https://www.rfc-editor.org/rfc/rfc3986#section-3.4">URI query</a> component
  *
  * <p>Notes:
-  <ul> <li>Sensitive content provided in query string SHOULD be scrubbed when instrumentations can
- identify it.</li> </ul>
+  <ul> <li>Sensitive content provided in {@code url.query} SHOULD be scrubbed when instrumentations
+ can identify it.</li> </ul>
  */
 static constexpr const char *kUrlQuery = "url.query";
+
+/**
+ * The highest registered url domain, stripped of the subdomain.
+ *
+ * <p>Notes:
+  <ul> <li>This value can be determined precisely with the <a href="http://publicsuffix.org">public
+ suffix list</a>. For example, the registered domain for {@code foo.example.com} is {@code
+ example.com}. Trying to approximate this by simply taking the last two labels will not work well
+ for TLDs such as {@code co.uk}.</li> </ul>
+ */
+static constexpr const char *kUrlRegisteredDomain = "url.registered_domain";
 
 /**
  * The <a href="https://www.rfc-editor.org/rfc/rfc3986#section-3.1">URI scheme</a> component
@@ -1515,20 +2728,55 @@ static constexpr const char *kUrlQuery = "url.query";
 static constexpr const char *kUrlScheme = "url.scheme";
 
 /**
+ * The subdomain portion of a fully qualified domain name includes all of the names except the host
+ name under the registered_domain. In a partially qualified domain, or if the qualification level of
+ the full name cannot be determined, subdomain contains all of the names below the registered
+ domain.
+ *
+ * <p>Notes:
+  <ul> <li>The subdomain portion of {@code www.east.mydomain.co.uk} is {@code east}. If the domain
+ has multiple levels of subdomain, such as {@code sub2.sub1.example.com}, the subdomain field should
+ contain {@code sub2.sub1}, with no trailing period.</li> </ul>
+ */
+static constexpr const char *kUrlSubdomain = "url.subdomain";
+
+/**
+ * The effective top level domain (eTLD), also known as the domain suffix, is the last part of the
+ domain name. For example, the top level domain for example.com is {@code com}.
+ *
+ * <p>Notes:
+  <ul> <li>This value can be determined precisely with the <a href="http://publicsuffix.org">public
+ suffix list</a>.</li> </ul>
+ */
+static constexpr const char *kUrlTopLevelDomain = "url.top_level_domain";
+
+/**
+ * Name of the user-agent extracted from original. Usually refers to the browser's name.
+ *
+ * <p>Notes:
+  <ul> <li><a href="https://www.whatsmyua.info">Example</a> of extracting browser's name from
+ original string. In the case of using a user-agent for non-browser products, such as microservices
+ with multiple names/versions inside the {@code user_agent.original}, the most significant name
+ SHOULD be selected. In such a scenario it should align with {@code user_agent.version}</li> </ul>
+ */
+static constexpr const char *kUserAgentName = "user_agent.name";
+
+/**
  * Value of the <a href="https://www.rfc-editor.org/rfc/rfc9110.html#field.user-agent">HTTP
  * User-Agent</a> header sent by the client.
  */
 static constexpr const char *kUserAgentOriginal = "user_agent.original";
 
 /**
- * A unique id to identify a session.
+ * Version of the user-agent extracted from original. Usually refers to the browser's version
+ *
+ * <p>Notes:
+  <ul> <li><a href="https://www.whatsmyua.info">Example</a> of extracting browser's version from
+ original string. In the case of using a user-agent for non-browser products, such as microservices
+ with multiple names/versions inside the {@code user_agent.original}, the most significant version
+ SHOULD be selected. In such a scenario it should align with {@code user_agent.name}</li> </ul>
  */
-static constexpr const char *kSessionId = "session.id";
-
-/**
- * The previous {@code session.id} for this user, when known.
- */
-static constexpr const char *kSessionPreviousId = "session.previous_id";
+static constexpr const char *kUserAgentVersion = "user_agent.version";
 
 /**
  * The full invoked ARN as provided on the {@code Context} passed to the function ({@code
@@ -1538,40 +2786,6 @@ static constexpr const char *kSessionPreviousId = "session.previous_id";
   <ul> <li>This may be different from {@code cloud.resource_id} if an alias is involved.</li> </ul>
  */
 static constexpr const char *kAwsLambdaInvokedArn = "aws.lambda.invoked_arn";
-
-/**
- * The <a href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#id">event_id</a>
- * uniquely identifies the event.
- */
-static constexpr const char *kCloudeventsEventId = "cloudevents.event_id";
-
-/**
- * The <a
- * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#source-1">source</a>
- * identifies the context in which an event happened.
- */
-static constexpr const char *kCloudeventsEventSource = "cloudevents.event_source";
-
-/**
- * The <a
- * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#specversion">version of
- * the CloudEvents specification</a> which the event uses.
- */
-static constexpr const char *kCloudeventsEventSpecVersion = "cloudevents.event_spec_version";
-
-/**
- * The <a
- * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#subject">subject</a> of
- * the event in the context of the event producer (identified by source).
- */
-static constexpr const char *kCloudeventsEventSubject = "cloudevents.event_subject";
-
-/**
- * The <a
- * href="https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md#type">event_type</a>
- * contains a value describing the type of event related to the originating occurrence.
- */
-static constexpr const char *kCloudeventsEventType = "cloudevents.event_type";
 
 /**
  * Parent-child Reference type
@@ -1593,201 +2807,10 @@ static constexpr const char *kOtelStatusCode = "otel.status_code";
 static constexpr const char *kOtelStatusDescription = "otel.status_description";
 
 /**
- * The invocation ID of the current function invocation.
- */
-static constexpr const char *kFaasInvocationId = "faas.invocation_id";
-
-/**
- * The name of the source on which the triggering operation was performed. For example, in Cloud
- * Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.
- */
-static constexpr const char *kFaasDocumentCollection = "faas.document.collection";
-
-/**
- * The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the
- * name of the file, and in Cosmos DB the table name.
- */
-static constexpr const char *kFaasDocumentName = "faas.document.name";
-
-/**
- * Describes the type of the operation that was performed on the data.
- */
-static constexpr const char *kFaasDocumentOperation = "faas.document.operation";
-
-/**
- * A string containing the time when the data was accessed in the <a
- * href="https://www.iso.org/iso-8601-date-and-time-format.html">ISO 8601</a> format expressed in <a
- * href="https://www.w3.org/TR/NOTE-datetime">UTC</a>.
- */
-static constexpr const char *kFaasDocumentTime = "faas.document.time";
-
-/**
- * A string containing the schedule period as <a
- * href="https://docs.oracle.com/cd/E12058_01/doc/doc.1014/e12030/cron_expressions.htm">Cron
- * Expression</a>.
- */
-static constexpr const char *kFaasCron = "faas.cron";
-
-/**
- * A string containing the function invocation time in the <a
- * href="https://www.iso.org/iso-8601-date-and-time-format.html">ISO 8601</a> format expressed in <a
- * href="https://www.w3.org/TR/NOTE-datetime">UTC</a>.
- */
-static constexpr const char *kFaasTime = "faas.time";
-
-/**
- * A boolean that is true if the serverless function is executed for the first time (aka
- * cold-start).
- */
-static constexpr const char *kFaasColdstart = "faas.coldstart";
-
-/**
- * The unique identifier of the feature flag.
- */
-static constexpr const char *kFeatureFlagKey = "feature_flag.key";
-
-/**
- * The name of the service provider that performs the flag evaluation.
- */
-static constexpr const char *kFeatureFlagProviderName = "feature_flag.provider_name";
-
-/**
- * SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the
-value can be used.
- *
- * <p>Notes:
-  <ul> <li>A semantic identifier, commonly referred to as a variant, provides a means
-for referring to a value without including the value itself. This can
-provide additional context for understanding the meaning behind a value.
-For example, the variant {@code red} maybe be used for the value {@code #c05543}.</li><li>A
-stringified version of the value can be used in situations where a semantic identifier is
-unavailable. String representation of the value should be determined by the implementer.</li> </ul>
- */
-static constexpr const char *kFeatureFlagVariant = "feature_flag.variant";
-
-/**
  * The AWS request ID as returned in the response headers {@code x-amz-request-id} or {@code
  * x-amz-requestid}.
  */
 static constexpr const char *kAwsRequestId = "aws.request_id";
-
-/**
- * The value of the {@code AttributesToGet} request parameter.
- */
-static constexpr const char *kAwsDynamodbAttributesToGet = "aws.dynamodb.attributes_to_get";
-
-/**
- * The value of the {@code ConsistentRead} request parameter.
- */
-static constexpr const char *kAwsDynamodbConsistentRead = "aws.dynamodb.consistent_read";
-
-/**
- * The JSON-serialized value of each item in the {@code ConsumedCapacity} response field.
- */
-static constexpr const char *kAwsDynamodbConsumedCapacity = "aws.dynamodb.consumed_capacity";
-
-/**
- * The value of the {@code IndexName} request parameter.
- */
-static constexpr const char *kAwsDynamodbIndexName = "aws.dynamodb.index_name";
-
-/**
- * The JSON-serialized value of the {@code ItemCollectionMetrics} response field.
- */
-static constexpr const char *kAwsDynamodbItemCollectionMetrics =
-    "aws.dynamodb.item_collection_metrics";
-
-/**
- * The value of the {@code Limit} request parameter.
- */
-static constexpr const char *kAwsDynamodbLimit = "aws.dynamodb.limit";
-
-/**
- * The value of the {@code ProjectionExpression} request parameter.
- */
-static constexpr const char *kAwsDynamodbProjection = "aws.dynamodb.projection";
-
-/**
- * The value of the {@code ProvisionedThroughput.ReadCapacityUnits} request parameter.
- */
-static constexpr const char *kAwsDynamodbProvisionedReadCapacity =
-    "aws.dynamodb.provisioned_read_capacity";
-
-/**
- * The value of the {@code ProvisionedThroughput.WriteCapacityUnits} request parameter.
- */
-static constexpr const char *kAwsDynamodbProvisionedWriteCapacity =
-    "aws.dynamodb.provisioned_write_capacity";
-
-/**
- * The value of the {@code Select} request parameter.
- */
-static constexpr const char *kAwsDynamodbSelect = "aws.dynamodb.select";
-
-/**
- * The keys in the {@code RequestItems} object field.
- */
-static constexpr const char *kAwsDynamodbTableNames = "aws.dynamodb.table_names";
-
-/**
- * The JSON-serialized value of each item of the {@code GlobalSecondaryIndexes} request field
- */
-static constexpr const char *kAwsDynamodbGlobalSecondaryIndexes =
-    "aws.dynamodb.global_secondary_indexes";
-
-/**
- * The JSON-serialized value of each item of the {@code LocalSecondaryIndexes} request field.
- */
-static constexpr const char *kAwsDynamodbLocalSecondaryIndexes =
-    "aws.dynamodb.local_secondary_indexes";
-
-/**
- * The value of the {@code ExclusiveStartTableName} request parameter.
- */
-static constexpr const char *kAwsDynamodbExclusiveStartTable = "aws.dynamodb.exclusive_start_table";
-
-/**
- * The the number of items in the {@code TableNames} response parameter.
- */
-static constexpr const char *kAwsDynamodbTableCount = "aws.dynamodb.table_count";
-
-/**
- * The value of the {@code ScanIndexForward} request parameter.
- */
-static constexpr const char *kAwsDynamodbScanForward = "aws.dynamodb.scan_forward";
-
-/**
- * The value of the {@code Count} response parameter.
- */
-static constexpr const char *kAwsDynamodbCount = "aws.dynamodb.count";
-
-/**
- * The value of the {@code ScannedCount} response parameter.
- */
-static constexpr const char *kAwsDynamodbScannedCount = "aws.dynamodb.scanned_count";
-
-/**
- * The value of the {@code Segment} request parameter.
- */
-static constexpr const char *kAwsDynamodbSegment = "aws.dynamodb.segment";
-
-/**
- * The value of the {@code TotalSegments} request parameter.
- */
-static constexpr const char *kAwsDynamodbTotalSegments = "aws.dynamodb.total_segments";
-
-/**
- * The JSON-serialized value of each item in the {@code AttributeDefinitions} request field.
- */
-static constexpr const char *kAwsDynamodbAttributeDefinitions =
-    "aws.dynamodb.attribute_definitions";
-
-/**
- * The JSON-serialized value of each item in the the {@code GlobalSecondaryIndexUpdates} request
- * field.
- */
-static constexpr const char *kAwsDynamodbGlobalSecondaryIndexUpdates =
-    "aws.dynamodb.global_secondary_index_updates";
 
 /**
  * The S3 bucket name the request refers to. Corresponds to the {@code --bucket} parameter of the <a
@@ -1944,34 +2967,6 @@ static constexpr const char *kMessageType = "message.type";
 static constexpr const char *kMessageUncompressedSize = "message.uncompressed_size";
 
 // Enum definitions
-namespace FaasInvokedProviderValues
-{
-/** Alibaba Cloud. */
-static constexpr const char *kAlibabaCloud = "alibaba_cloud";
-/** Amazon Web Services. */
-static constexpr const char *kAws = "aws";
-/** Microsoft Azure. */
-static constexpr const char *kAzure = "azure";
-/** Google Cloud Platform. */
-static constexpr const char *kGcp = "gcp";
-/** Tencent Cloud. */
-static constexpr const char *kTencentCloud = "tencent_cloud";
-}  // namespace FaasInvokedProviderValues
-
-namespace FaasTriggerValues
-{
-/** A response to some data source operation such as a database or filesystem read/write. */
-static constexpr const char *kDatasource = "datasource";
-/** To provide an answer to an inbound HTTP request. */
-static constexpr const char *kHttp = "http";
-/** A function is set to be executed when messages are sent to a messaging system. */
-static constexpr const char *kPubsub = "pubsub";
-/** A function is scheduled to be executed regularly. */
-static constexpr const char *kTimer = "timer";
-/** If none of the others apply. */
-static constexpr const char *kOther = "other";
-}  // namespace FaasTriggerValues
-
 namespace LogIostreamValues
 {
 /** Logs from stdout stream. */
@@ -2056,6 +3051,16 @@ static constexpr const char *kHeap = "heap";
 /** Non-heap memory. */
 static constexpr const char *kNonHeap = "non_heap";
 }  // namespace JvmMemoryTypeValues
+
+namespace ProcessCpuStateValues
+{
+/** system. */
+static constexpr const char *kSystem = "system";
+/** user. */
+static constexpr const char *kUser = "user";
+/** wait. */
+static constexpr const char *kWait = "wait";
+}  // namespace ProcessCpuStateValues
 
 namespace SystemCpuStateValues
 {
@@ -2167,7 +3172,7 @@ static constexpr const char *kSynSent = "syn_sent";
 static constexpr const char *kTimeWait = "time_wait";
 }  // namespace SystemNetworkStateValues
 
-namespace SystemProcessesStatusValues
+namespace SystemProcessStatusValues
 {
 /** running. */
 static constexpr const char *kRunning = "running";
@@ -2177,7 +3182,97 @@ static constexpr const char *kSleeping = "sleeping";
 static constexpr const char *kStopped = "stopped";
 /** defunct. */
 static constexpr const char *kDefunct = "defunct";
-}  // namespace SystemProcessesStatusValues
+}  // namespace SystemProcessStatusValues
+
+namespace CloudPlatformValues
+{
+/** Alibaba Cloud Elastic Compute Service. */
+static constexpr const char *kAlibabaCloudEcs = "alibaba_cloud_ecs";
+/** Alibaba Cloud Function Compute. */
+static constexpr const char *kAlibabaCloudFc = "alibaba_cloud_fc";
+/** Red Hat OpenShift on Alibaba Cloud. */
+static constexpr const char *kAlibabaCloudOpenshift = "alibaba_cloud_openshift";
+/** AWS Elastic Compute Cloud. */
+static constexpr const char *kAwsEc2 = "aws_ec2";
+/** AWS Elastic Container Service. */
+static constexpr const char *kAwsEcs = "aws_ecs";
+/** AWS Elastic Kubernetes Service. */
+static constexpr const char *kAwsEks = "aws_eks";
+/** AWS Lambda. */
+static constexpr const char *kAwsLambda = "aws_lambda";
+/** AWS Elastic Beanstalk. */
+static constexpr const char *kAwsElasticBeanstalk = "aws_elastic_beanstalk";
+/** AWS App Runner. */
+static constexpr const char *kAwsAppRunner = "aws_app_runner";
+/** Red Hat OpenShift on AWS (ROSA). */
+static constexpr const char *kAwsOpenshift = "aws_openshift";
+/** Azure Virtual Machines. */
+static constexpr const char *kAzureVm = "azure_vm";
+/** Azure Container Apps. */
+static constexpr const char *kAzureContainerApps = "azure_container_apps";
+/** Azure Container Instances. */
+static constexpr const char *kAzureContainerInstances = "azure_container_instances";
+/** Azure Kubernetes Service. */
+static constexpr const char *kAzureAks = "azure_aks";
+/** Azure Functions. */
+static constexpr const char *kAzureFunctions = "azure_functions";
+/** Azure App Service. */
+static constexpr const char *kAzureAppService = "azure_app_service";
+/** Azure Red Hat OpenShift. */
+static constexpr const char *kAzureOpenshift = "azure_openshift";
+/** Google Bare Metal Solution (BMS). */
+static constexpr const char *kGcpBareMetalSolution = "gcp_bare_metal_solution";
+/** Google Cloud Compute Engine (GCE). */
+static constexpr const char *kGcpComputeEngine = "gcp_compute_engine";
+/** Google Cloud Run. */
+static constexpr const char *kGcpCloudRun = "gcp_cloud_run";
+/** Google Cloud Kubernetes Engine (GKE). */
+static constexpr const char *kGcpKubernetesEngine = "gcp_kubernetes_engine";
+/** Google Cloud Functions (GCF). */
+static constexpr const char *kGcpCloudFunctions = "gcp_cloud_functions";
+/** Google Cloud App Engine (GAE). */
+static constexpr const char *kGcpAppEngine = "gcp_app_engine";
+/** Red Hat OpenShift on Google Cloud. */
+static constexpr const char *kGcpOpenshift = "gcp_openshift";
+/** Red Hat OpenShift on IBM Cloud. */
+static constexpr const char *kIbmCloudOpenshift = "ibm_cloud_openshift";
+/** Tencent Cloud Cloud Virtual Machine (CVM). */
+static constexpr const char *kTencentCloudCvm = "tencent_cloud_cvm";
+/** Tencent Cloud Elastic Kubernetes Service (EKS). */
+static constexpr const char *kTencentCloudEks = "tencent_cloud_eks";
+/** Tencent Cloud Serverless Cloud Function (SCF). */
+static constexpr const char *kTencentCloudScf = "tencent_cloud_scf";
+}  // namespace CloudPlatformValues
+
+namespace CloudProviderValues
+{
+/** Alibaba Cloud. */
+static constexpr const char *kAlibabaCloud = "alibaba_cloud";
+/** Amazon Web Services. */
+static constexpr const char *kAws = "aws";
+/** Microsoft Azure. */
+static constexpr const char *kAzure = "azure";
+/** Google Cloud Platform. */
+static constexpr const char *kGcp = "gcp";
+/** Heroku Platform as a Service. */
+static constexpr const char *kHeroku = "heroku";
+/** IBM Cloud. */
+static constexpr const char *kIbmCloud = "ibm_cloud";
+/** Tencent Cloud. */
+static constexpr const char *kTencentCloud = "tencent_cloud";
+}  // namespace CloudProviderValues
+
+namespace ContainerCpuStateValues
+{
+/** When tasks of the cgroup are in user mode (Linux). When all container processes are in user mode
+ * (Windows). */
+static constexpr const char *kUser = "user";
+/** When CPU is used by the system (host OS). */
+static constexpr const char *kSystem = "system";
+/** When tasks of the cgroup are in kernel mode (Linux). When all container processes are in kernel
+ * mode (Windows). */
+static constexpr const char *kKernel = "kernel";
+}  // namespace ContainerCpuStateValues
 
 namespace DbCassandraConsistencyLevelValues
 {
@@ -2395,6 +3490,18 @@ static constexpr const char *kInproc = "inproc";
 static constexpr const char *kOther = "other";
 }  // namespace NetTransportValues
 
+namespace SystemProcessesStatusValues
+{
+/** running. */
+static constexpr const char *kRunning = "running";
+/** sleeping. */
+static constexpr const char *kSleeping = "sleeping";
+/** stopped. */
+static constexpr const char *kStopped = "stopped";
+/** defunct. */
+static constexpr const char *kDefunct = "defunct";
+}  // namespace SystemProcessesStatusValues
+
 namespace DiskIoDirectionValues
 {
 /** read. */
@@ -2408,6 +3515,72 @@ namespace ErrorTypeValues
 /** A fallback error value to be used when the instrumentation doesn&#39;t define a custom value. */
 static constexpr const char *kOther = "_OTHER";
 }  // namespace ErrorTypeValues
+
+namespace FaasDocumentOperationValues
+{
+/** When a new object is created. */
+static constexpr const char *kInsert = "insert";
+/** When an object is modified. */
+static constexpr const char *kEdit = "edit";
+/** When an object is deleted. */
+static constexpr const char *kDelete = "delete";
+}  // namespace FaasDocumentOperationValues
+
+namespace FaasInvokedProviderValues
+{
+/** Alibaba Cloud. */
+static constexpr const char *kAlibabaCloud = "alibaba_cloud";
+/** Amazon Web Services. */
+static constexpr const char *kAws = "aws";
+/** Microsoft Azure. */
+static constexpr const char *kAzure = "azure";
+/** Google Cloud Platform. */
+static constexpr const char *kGcp = "gcp";
+/** Tencent Cloud. */
+static constexpr const char *kTencentCloud = "tencent_cloud";
+}  // namespace FaasInvokedProviderValues
+
+namespace FaasTriggerValues
+{
+/** A response to some data source operation such as a database or filesystem read/write. */
+static constexpr const char *kDatasource = "datasource";
+/** To provide an answer to an inbound HTTP request. */
+static constexpr const char *kHttp = "http";
+/** A function is set to be executed when messages are sent to a messaging system. */
+static constexpr const char *kPubsub = "pubsub";
+/** A function is scheduled to be executed regularly. */
+static constexpr const char *kTimer = "timer";
+/** If none of the others apply. */
+static constexpr const char *kOther = "other";
+}  // namespace FaasTriggerValues
+
+namespace HostArchValues
+{
+/** AMD64. */
+static constexpr const char *kAmd64 = "amd64";
+/** ARM32. */
+static constexpr const char *kArm32 = "arm32";
+/** ARM64. */
+static constexpr const char *kArm64 = "arm64";
+/** Itanium. */
+static constexpr const char *kIa64 = "ia64";
+/** 32-bit PowerPC. */
+static constexpr const char *kPpc32 = "ppc32";
+/** 64-bit PowerPC. */
+static constexpr const char *kPpc64 = "ppc64";
+/** IBM z/Architecture. */
+static constexpr const char *kS390x = "s390x";
+/** 32-bit x86. */
+static constexpr const char *kX86 = "x86";
+}  // namespace HostArchValues
+
+namespace HttpConnectionStateValues
+{
+/** active state. */
+static constexpr const char *kActive = "active";
+/** idle state. */
+static constexpr const char *kIdle = "idle";
+}  // namespace HttpConnectionStateValues
 
 namespace HttpRequestMethodValues
 {
@@ -2445,9 +3618,10 @@ static constexpr const char *kCreate = "create";
 /** One or more messages are requested by a consumer. This operation refers to pull-based scenarios,
  * where consumers explicitly call methods of messaging SDKs to receive messages. */
 static constexpr const char *kReceive = "receive";
-/** One or more messages are passed to a consumer. This operation refers to push-based scenarios,
- * where consumer register callbacks which get called by messaging SDKs. */
-static constexpr const char *kDeliver = "deliver";
+/** One or more messages are delivered to or processed by a consumer. */
+static constexpr const char *kDeliver = "process";
+/** One or more messages are settled. */
+static constexpr const char *kSettle = "settle";
 }  // namespace MessagingOperationValues
 
 namespace MessagingRocketmqConsumptionModelValues
@@ -2470,6 +3644,18 @@ static constexpr const char *kDelay = "delay";
 static constexpr const char *kTransaction = "transaction";
 }  // namespace MessagingRocketmqMessageTypeValues
 
+namespace MessagingServicebusDispositionStatusValues
+{
+/** Message is completed. */
+static constexpr const char *kComplete = "complete";
+/** Message is abandoned. */
+static constexpr const char *kAbandon = "abandon";
+/** Message is sent to dead letter queue. */
+static constexpr const char *kDeadLetter = "dead_letter";
+/** Message is deferred. */
+static constexpr const char *kDefer = "defer";
+}  // namespace MessagingServicebusDispositionStatusValues
+
 namespace MessagingSystemValues
 {
 /** Apache ActiveMQ. */
@@ -2477,11 +3663,11 @@ static constexpr const char *kActivemq = "activemq";
 /** Amazon Simple Queue Service (SQS). */
 static constexpr const char *kAwsSqs = "aws_sqs";
 /** Azure Event Grid. */
-static constexpr const char *kAzureEventgrid = "azure_eventgrid";
+static constexpr const char *kEventgrid = "eventgrid";
 /** Azure Event Hubs. */
-static constexpr const char *kAzureEventhubs = "azure_eventhubs";
+static constexpr const char *kEventhubs = "eventhubs";
 /** Azure Service Bus. */
-static constexpr const char *kAzureServicebus = "azure_servicebus";
+static constexpr const char *kServicebus = "servicebus";
 /** Google Cloud Pub/Sub. */
 static constexpr const char *kGcpPubsub = "gcp_pubsub";
 /** Java Message Service. */
@@ -2582,6 +3768,32 @@ static constexpr const char *kIpv4 = "ipv4";
 static constexpr const char *kIpv6 = "ipv6";
 }  // namespace NetworkTypeValues
 
+namespace OsTypeValues
+{
+/** Microsoft Windows. */
+static constexpr const char *kWindows = "windows";
+/** Linux. */
+static constexpr const char *kLinux = "linux";
+/** Apple Darwin. */
+static constexpr const char *kDarwin = "darwin";
+/** FreeBSD. */
+static constexpr const char *kFreebsd = "freebsd";
+/** NetBSD. */
+static constexpr const char *kNetbsd = "netbsd";
+/** OpenBSD. */
+static constexpr const char *kOpenbsd = "openbsd";
+/** DragonFly BSD. */
+static constexpr const char *kDragonflybsd = "dragonflybsd";
+/** HP-UX (Hewlett Packard Unix). */
+static constexpr const char *kHpux = "hpux";
+/** AIX (Advanced Interactive eXecutive). */
+static constexpr const char *kAix = "aix";
+/** SunOS, Oracle Solaris. */
+static constexpr const char *kSolaris = "solaris";
+/** IBM z/OS. */
+static constexpr const char *kZOs = "z_os";
+}  // namespace OsTypeValues
+
 namespace RpcConnectRpcErrorCodeValues
 {
 /** cancelled. */
@@ -2670,6 +3882,34 @@ static constexpr const char *kApacheDubbo = "apache_dubbo";
 static constexpr const char *kConnectRpc = "connect_rpc";
 }  // namespace RpcSystemValues
 
+namespace TelemetrySdkLanguageValues
+{
+/** cpp. */
+static constexpr const char *kCpp = "cpp";
+/** dotnet. */
+static constexpr const char *kDotnet = "dotnet";
+/** erlang. */
+static constexpr const char *kErlang = "erlang";
+/** go. */
+static constexpr const char *kGo = "go";
+/** java. */
+static constexpr const char *kJava = "java";
+/** nodejs. */
+static constexpr const char *kNodejs = "nodejs";
+/** php. */
+static constexpr const char *kPhp = "php";
+/** python. */
+static constexpr const char *kPython = "python";
+/** ruby. */
+static constexpr const char *kRuby = "ruby";
+/** rust. */
+static constexpr const char *kRust = "rust";
+/** swift. */
+static constexpr const char *kSwift = "swift";
+/** webjs. */
+static constexpr const char *kWebjs = "webjs";
+}  // namespace TelemetrySdkLanguageValues
+
 namespace TlsProtocolNameValues
 {
 /** ssl. */
@@ -2694,16 +3934,6 @@ static constexpr const char *kOk = "OK";
 /** The operation contains an error. */
 static constexpr const char *kError = "ERROR";
 }  // namespace OtelStatusCodeValues
-
-namespace FaasDocumentOperationValues
-{
-/** When a new object is created. */
-static constexpr const char *kInsert = "insert";
-/** When an object is modified. */
-static constexpr const char *kEdit = "edit";
-/** When an object is deleted. */
-static constexpr const char *kDelete = "delete";
-}  // namespace FaasDocumentOperationValues
 
 namespace GraphqlOperationTypeValues
 {
